pyproject.toml
[tool.poetry]
name = "python-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.6"
djangorestframework = "3.15.2"
django-filter = "24.3"
django-debug-toolbar = "4.4.6"
pillow = "10.3.0"
gunicorn = "23.0.0"
asgiref = "3.8.1"
attrs = "24.2.0"
certifi = "2024.8.30"
docutils = "0.21.2"
drf-spectacular = "0.27.2"
flake8 = "7.1.1"
flake8-docstrings = "1.7.0"
inflection = "0.5.1"
jsonschema = "4.23.0"
jsonschema-specifications = "2023.12.1"
mccabe = "0.7.0"
packaging = "24.1"
pycodestyle = "2.12.1"
pydocstyle = "6.3.0"
pyflakes = "3.2.0"
pyyaml = "6.0.2"
referencing = "0.35.1"
rpds-py = "0.20.0"
sentry-sdk = "2.13.0"
snowballstemmer = "2.2.0"
sqlparse = "0.5.0"
typing-extensions = "4.11.0"
uritemplate = "4.1.1"
urllib3 = "2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
